HostDisplay: Common texture base class for all APIs

+++ src/duckstation-nogui/win32_nogui_platform.cpp
@@ -327,8 +327,8 @@ LRESULT CALLBACK Win32NoGUIPlatform::WndProc(HWND hwnd, UINT msg, WPARAM wParam,
       {
         const WCHAR utf16[1] = {static_cast<WCHAR>(wParam)};
         char utf8[8] = {};
-        const int utf8_len =
-          WideCharToMultiByte(CP_UTF8, 0, utf16, std::size(utf16), utf8, sizeof(utf8) - 1, nullptr, nullptr);
+        const int utf8_len = WideCharToMultiByte(CP_UTF8, 0, utf16, static_cast<int>(std::size(utf16)), utf8,
+                                                 static_cast<int>(sizeof(utf8)) - 1, nullptr, nullptr);
         if (utf8_len > 0)
         {
           utf8[utf8_len] = 0;
